+++ apiserver/grpcserver/base_test.go
@@ -24,7 +24,9 @@ import (
 
 	"google.golang.org/grpc/metadata"
 
+	"github.com/polarismesh/polaris/apiserver"
 	"github.com/polarismesh/polaris/common/utils"
+	grpchelp "github.com/polarismesh/polaris/apiserver/grpcserver/utils"
 )
 
 func mockGrpcContext(testVal map[string]string) context.Context {
@@ -116,6 +118,60 @@ func TestConvertContext(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := ConvertContext(tt.args.ctx); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("ConvertContext() = %v, \n want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestGetClientOpenMethod(t *testing.T) {
+	type args struct {
+		include  []string
+		protocol string
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    map[string]bool
+		wantErr bool
+	}{
+		{
+			name: "case=1",
+			args: args{
+				include: []string{
+					apiserver.RegisterAccess,
+				},
+				protocol: "grpc",
+			},
+			want: map[string]bool{
+				"/v1.PolarisGRPC/RegisterInstance":   true,
+				"/v1.PolarisGRPC/DeregisterInstance": true,
+			},
+			wantErr: false,
+		},
+		{
+			name: "case=1",
+			args: args{
+				include: []string{
+					apiserver.DiscoverAccess,
+				},
+				protocol: "grpc",
+			},
+			want: map[string]bool{
+				"/v1.PolarisGRPC/Discover":     true,
+				"/v1.PolarisGRPC/ReportClient": true,
+			},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := grpchelp.GetClientOpenMethod(tt.args.include, tt.args.protocol)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("GetClientOpenMethod() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("GetClientOpenMethod() = %v, want %v", got, tt.want)
 			}
 		})
 	}
